requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
langchain
langchain-openai
langchain-community
langchain-core
supabase
openai
python-multipart
